Rename ValveConfigurationAndControl cluster files
