+++ src/app/clusters/valve-configuration-and-control-server/app_config_dependent_sources.gni
@@ -13,6 +13,6 @@
 # limitations under the License.
 app_config_dependent_sources = [
   "valve-configuration-and-control-delegate.h",
-  "valve-configuration-and-control-server.cpp",
-  "valve-configuration-and-control-server.h",
+  "valve-configuration-and-control-cluster.cpp",
+  "valve-configuration-and-control-cluster.h",
 ]
